Use AttachmentFragment type in attachments.jsx
